requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
pyspark==3.4.1

# Database connectivity
sqlalchemy==1.4.46
pymssql==2.2.7

# Data processing and validation
pydantic==1.10.8
python-dateutil==2.8.2

# HTTP requests for API calls
requests==2.28.2

# Configuration and logging
pyyaml==6.0
python-dotenv==1.0.0
structlog==23.2.0

# Utilities
tqdm==4.64.1
colorama==0.4.6

# Development and testing
pytest==7.3.1
pytest-cov==4.0.0 